// src/education/users/item/assignments/item/resources/item/educationAssignmentResourceItemRequestBuilderPatchRequestConfiguration.ts
import type {RequestOption} from '@microsoft/kiota-abstractions';

export interface EducationAssignmentResourceItemRequestBuilderPatchRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}